Refactor the code for bulk upload to different function
