+++ src/index.ts
@@ -1,29 +1,14 @@
 import fs from 'fs';
 import csv from 'csv-parser';
-import { CHA_TABLES, DATA_FILE_PATH } from './config/config';
+import { DATA_FILE_PATH } from './config/config';
 
 import { getCSRFToken, getFormattedHeaders, loginResult } from './utils/auth';
 
-import {
-  getRoles,
-  generateRole,
-  generatePermissions,
-  createUserRole,
-  SupersetRole,
-} from './utils/role';
+import { getRoles } from './utils/role';
 
-import {
-  createRowlevelSecurity,
-  generateRowLevelSecurity,
-  getAvailableRowlevelSecurityFromSuperset,
-} from './utils/rowlevelsecurity';
-import { CSVUser, createUserAccount, generateUser } from './utils/user';
+import { CSVUser } from './utils/user';
 
-import { IRowLevelSecurityFromSuperset } from './utils/interface';
-import {
-  addPermissionsForUserRole,
-  getUserPermissions,
-} from './utils/permissions';
+import { bulkUserUploadUg } from './bulk-user-upload';
 
 const readAndParse = async (fileName: string) => {
   const tokens = await loginResult();
@@ -38,15 +23,7 @@ const readAndParse = async (fileName: string) => {
 
   const rolesAvailableOnSuperset = await getRoles(headers);
 
-  const userPermissions = await getUserPermissions(
-    rolesAvailableOnSuperset,
-    headers,
-  );
-
   const users: CSVUser[] = [];
-
-  const { result: rowLevelFromSuperset } =
-    await getAvailableRowlevelSecurityFromSuperset(headers);
 
   fs.createReadStream(fileName, 'utf-8')
     .on('error', () => {
@@ -58,52 +35,7 @@ const readAndParse = async (fileName: string) => {
     .on('end', async () => {
       console.log(users);
       console.log(`Processed ${users.length} successfully`);
-
-      users.forEach(async (user) => {
-        let userRole: SupersetRole;
-
-        const generatedRole = generateRole(user.role, user.place);
-
-        const existingRoleOnSuperset = rolesAvailableOnSuperset.find(
-          (ssrole: { id: number; name: string }) =>
-            ssrole.name === generatedRole.name,
-        );
-
-        if (existingRoleOnSuperset) {
-          userRole = existingRoleOnSuperset;
-        } else {
-          userRole = await createUserRole(generatedRole, headers);
-          rolesAvailableOnSuperset.push({
-            id: userRole.id,
-            name: userRole.name,
-          });
-        }
-
-        const rolePermissions = generatePermissions(userPermissions);
-        await addPermissionsForUserRole(userRole.id, rolePermissions, headers);
-
-        const generatedUser = generateUser(user, [userRole.id]);
-        await createUserAccount(generatedUser, headers);
-
-        const rowLevelSecurity = generateRowLevelSecurity(
-          [userRole.id],
-          user.group,
-          user.place,
-          CHA_TABLES,
-          user.role,
-        );
-        const doesRowLevelExist = rowLevelFromSuperset.some(
-          (level: IRowLevelSecurityFromSuperset) =>
-            level.name === rowLevelSecurity.name,
-        );
-        if (!doesRowLevelExist) {
-          const response = await createRowlevelSecurity(
-            rowLevelSecurity,
-            headers,
-          );
-          rowLevelFromSuperset.push(response);
-        }
-      });
+      await bulkUserUploadUg(users, rolesAvailableOnSuperset, headers);
     });
 };
